requirements.txt
mcp
azure-identity
openai
pytest
pytest-asyncio
python-dotenv
